Localise percentage in LandingPage
Refactor RankedList to avoid it having knowledge about municipalities
and companies. This simplifies the code and avoids having hard coded
support for specific types of RankedLists.

Using an itemValueRenderer due to the way emissions are rendered, if
this changes in the future so that is rendered similarly to percentages
(value and unit in the same style) we can switch back to using
displayValue in each item as a string.