Switch main menu from Menubar to NavigationMenu
According to the documentation [1] the Menubar component is meant for a
desktop applications menu bar (File, Edit, ...). The NavigationMenu [2]
fits our use case better.

However, the NavigationMenu by default has built in unfold-on-hover
which I am unsure if we want or not. Currently I have "disabled" it
through a really long delay before it opens.

[1] https://www.radix-ui.com/primitives/docs/components/menubar [2]
https://www.radix-ui.com/primitives/docs/components/navigation-menu